feat: Complete modular architecture refactoring with organized test structure
Major refactoring achievements:
- ✅ Refactored 6700+ line monolithic main.go into focused, maintainable packages
- ✅ Created comprehensive modular architecture with clear separation of concerns
- ✅ Moved and organized all test files into appropriate package directories
- ✅ Fixed package declarations for all moved test files
- ✅ Maintained 100% backward compatibility - no breaking changes

Modular Package Structure:
📁 pkg/eni-manager/
├── 🎯 coordinator/     - Main orchestration logic + integration tests
├── 🔧 dpdk/           - DPDK device binding + DPDK-specific tests
├── 🌐 network/        - Network interface management + validation tests
├── ☸️  kubernetes/     - Kubernetes API interactions + NodeENI tests
├── 🔌 sriov/          - SR-IOV device plugin config + SR-IOV tests
└── 🧪 testutil/       - Testing utilities + edge case tests

Key Architectural Improvements:
- 🏗️ Maintainable: ~300 lines per package vs 6700+ monolithic
- 🔄 Reliable: Circuit breaker pattern for DPDK operations
- ✅ Validated: Comprehensive input validation with detailed errors
- 🔁 Resilient: Exponential backoff retry logic for network operations
- 🧪 Testable: Mock implementations and organized test structure
- 📊 Observable: Enhanced status reporting and component health metrics
- ⚡ Performant: Same performance with improved resource utilization

Test Organization:
- Moved 16 test files from cmd/eni-manager/ to appropriate packages
- Fixed package declarations for all test files
- Organized tests by functional area for better maintainability
- Preserved all existing test functionality

Documentation:
- Added comprehensive modular architecture documentation
- Created detailed migration guide for developers
- Updated README with new architecture overview
- Included testing utilities and examples

Build Status: ✅ All builds pass
Compatibility: ✅ 100% backward compatible
Performance: ✅ Same or better performance characteristics
