+++ pkg/eni-manager/coordinator/startup_force_restart_test.go
@@ -1,4 +1,4 @@
-package main
+package coordinator
 
 import (
 	"encoding/json"
