+++ pkg/eni-manager/dpdk/dpdk_driver_conflict_test.go
@@ -1,4 +1,4 @@
-package main
+package dpdk
 
 import (
 	"encoding/json"
