+++ pkg/eni-manager/dpdk/dpdk_operations_test.go
@@ -1,4 +1,4 @@
-package main
+package dpdk
 
 import (
 	"fmt"
