+++ pkg/eni-manager/dpdk/dpdk_verification_fix_test.go
@@ -1,4 +1,4 @@
-package main
+package dpdk
 
 import (
 	"regexp"
