+++ pkg/eni-manager/kubernetes/nodeeni_deletion_test.go
@@ -1,4 +1,4 @@
-package main
+package kubernetes
 
 import (
 	"testing"
