+++ pkg/eni-manager/sriov/enhanced_sriov_restart_test.go
@@ -1,4 +1,4 @@
-package main
+package sriov
 
 import (
 	"encoding/json"
