+++ pkg/eni-manager/sriov/sriov_restart_cloud_native_test.go
@@ -1,4 +1,4 @@
-package main
+package sriov
 
 import (
 	"testing"
