+++ pkg/eni-manager/sriov/sriov_restart_fix_test.go
@@ -1,4 +1,4 @@
-package main
+package sriov
 
 import (
 	"encoding/json"
