+++ pkg/eni-manager/sriov/sriov_tracking_test.go
@@ -1,4 +1,4 @@
-package main
+package sriov
 
 import (
 	"os"
