+++ pkg/eni-manager/testutil/cleanup_failure_scenarios_test.go
@@ -1,4 +1,4 @@
-package main
+package testutil
 
 import (
 	"fmt"
