+++ pkg/eni-manager/testutil/edge_cases_test.go
@@ -1,4 +1,4 @@
-package main
+package testutil
 
 import (
 	"encoding/json"
